last commit

+++ ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: rnomura <rnomura@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/01 23:24:49 by rnomura           #+#    #+#             */
-/*   Updated: 2024/05/08 15:00:58 by rnomura          ###   ########.fr       */
+/*   Updated: 2024/05/12 10:37:07 by rnomura          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static size_t	count_digit(long num)
+static size_t	count_digit(long long num)
 {
 	size_t	count;
 
@@ -27,7 +27,8 @@ static size_t	count_digit(long num)
 	return (count);
 }
 
-static void	getdigitstoarray(char *arr, size_t count, size_t sign, long num)
+static void	getdigitstoarray(char *arr, size_t count, size_t sign,
+		long long num)
 {
 	arr[count] = '\0';
 	while (sign < count)
@@ -41,10 +42,10 @@ static void	getdigitstoarray(char *arr, size_t count, size_t sign, long num)
 
 char	*ft_itoa(int n)
 {
-	size_t	sign;
-	size_t	count;
-	long	num;
-	char	*arr;
+	size_t		sign;
+	size_t		count;
+	long long	num;
+	char		*arr;
 
 	count = 0;
 	sign = 0;
